fix(auth): resolve 401 sync error by fixing audience token and enforcing strict PKCE flow
